% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_report.R
\name{ga_report}
\alias{ga_report}
\title{Google Analytics daily report}
\usage{
ga_report(track_url = NULL, auto_save = TRUE)
}
\arguments{
\item{track_url}{character. ex. http://qiita.com}

\item{auto_save}{logical.}

\item{...}{other function arguments.}
}
\description{
Google Analytics daily report
}

