% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notice_new_rpkg.R
\name{notice_new_rpkg}
\alias{notice_new_rpkg}
\title{Notice new released r-packages.}
\usage{
notice_new_rpkg(path2auth = ".httr-oauth",
  path2gmailr_json = "auth_gmailr.json", address2sent = "<address_to_sent>",
  address_from = "<address_from>")
}
\arguments{
\item{path2auth}{path to .httr-oauth}

\item{path2gmailr_json}{path to google auth json file}

\item{address2sent}{character.}

\item{address_from}{character. ex. Uryu Shinya <suika1127@gmail.com>}
}
\description{
Notice new released r-packages.
}
\examples{
\dontrun{
notice_new_rpkg(".httr-oauth", 
                "auth_gmailr.json", 
                "u_suika.30700@m.evernote.com", "
                 Uryu Shinya <suika1127@gmail.com>")
}
}

