% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dl_log.R
\name{pkg_dl_log}
\alias{pkg_dl_log}
\title{CRAN Mirror daily download log.}
\usage{
pkg_dl_log(path2token = "googlesheets_token.rds",
  path2conf = "slackr_setup")
}
\arguments{
\item{path2token}{character.}

\item{path2conf}{character.}
}
\description{
CRAN Mirror daily download log.
}
\examples{
\dontrun{
pkg_dl_log("googlesheets_token.rds", "slackr_setup")
}
}

